resolved style cop warnings
