+++ src/app/Model/ActorMovie.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System.Collections.Generic;
+
+namespace CSE.Helium.Model
+{
+    public class ActorMovie
+    {
+        public string MovieId { get; set; }
+        public string Title { get; set; }
+        public int Year { get; set; }
+        public int Runtime { get; set; }
+        public List<string> Genres { get; set; }
+    }
+}
